refactor: move inner class Point to separete class
Now we'll be able to create instances of the Point class as for any normal class.

Whilst the class could have been made static, if the program ever were to be extended it might make sense for other classes to also use the Point class.
