+++ decide/src/main/java/org/example/InitialSettings.java
@@ -59,22 +59,4 @@ public class InitialSettings {
             AREA2 = area2;
         }
     }
-
-    public class Point {
-        public final int x;
-        public final int y;
-
-        public Point(int x, int y) {
-            this.x = x;
-            this.y = y;
-        }
-
-        public double distance(Point other) {
-            return Math.sqrt(Math.pow(x - other.x, 2) + Math.pow(y - other.y, 2));
-        }
-
-        public static double triangleArea(Point p1, Point p2, Point p3) {
-            return Math.abs(p1.x * (p2.y - p3.y) + p2.x * (p3.y - p1.y) + p3.x * (p1.y - p2.y)) / 2.0;
-        }
-    }
 }
